Refactor express functionality into server.js
